requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.30.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0
audiomentations>=0.43.0
pydub>=0.25.1
scipy>=1.10.0
resampy>=0.4.2

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
h5py>=3.8.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Model Training & Evaluation
tensorboard>=2.13.0
wandb>=0.15.0
timm>=0.9.0  # PyTorch Image Models (for vision transformers)
einops>=0.6.1  # Tensor operations

# Metrics & Evaluation
torchmetrics>=1.0.0
scikit-metrics>=0.1.0

# Configuration Management
pyyaml>=6.0
omegaconf>=2.3.0
python-dotenv>=1.0.0

# Utilities
tqdm>=4.65.0
colorama>=0.4.6
rich>=13.4.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality
black>=23.3.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.4.0

# Jupyter
jupyter>=1.0.0
ipywidgets>=8.0.0
notebook>=7.0.0

# Additional Audio Tools
pyaudio>=0.2.13  # For real-time audio capture
webrtcvad>=2.0.10  # Voice activity detection
noisereduce>=2.0.0  # Noise reduction
